Exception handling - delete

+++ src/main/java/com/webproject/course/services/UserService.java
@@ -1,13 +1,13 @@
 package com.webproject.course.services;
 
-
 import com.webproject.course.entities.User;
 import com.webproject.course.repositories.UserRepository;
+import com.webproject.course.services.exceptions.DataBaseException;
 import com.webproject.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.Optional;
 
@@ -29,9 +29,19 @@ public class UserService{
     public User insert(User obj){
         return userRepository.save(obj);
     }
+
     public void delete(Long id){
-        userRepository.deleteById(id);
+        if(!userRepository.existsById(id)){
+            throw new ResourceNotFoundException(id);
+        }
+
+        try{
+            userRepository.deleteById(id);
+        }catch (DataIntegrityViolationException e ){
+            throw new DataBaseException(e.getMessage());
+        }
     }
+
     public User update(Long id, User obj){
         User entity = userRepository.getReferenceById(id);
         updateData(entity, obj);
